+++ Changes.md
@@ -10,11 +10,19 @@ Improvements
 ------------
 
 - Wrapper : Added warning when GafferRenderMan is not available for current RenderMan version.
+- Spreadsheet : A wider range of types are converted when copy/pasting values between cells, such as BoolData to IntData, FloatData to IntData, etc.
 
 Fixes
 -----
 
 - RenderMan : Fixed bug preventing startup files from being loaded from versioned GafferRenderMan modules.
+
+API
+---
+
+- PlugAlgo :
+  - Added support in `setValueFromData()` for setting StringPlug values from StringVectorData and StringVectorDataPlugValues from StringData.
+  - Added `setValueOrAddKeyFromData()`.
 
 1.5.9.0 (relative to 1.5.8.0)
 =======